+++ pages/account/settings/index.tsx
@@ -1,6 +1,6 @@
 import { Box, Container } from '@chakra-ui/react'
 import { GetServerSideProps } from 'next'
-import { SettingsTemplate } from '../../../src/components/account/settings/SettingTemplate'
+import { SettingsTemplate } from '@components/account/settings/SettingTemplate'
 const SettingsPage = () => {
   return (
     <Box w='full' h='full' bg='gray.100'>
