+++ src/components/account/settings/SettingTemplate.tsx
@@ -1,9 +1,10 @@
-import { Box, Flex, useDisclosure, VStack } from '@chakra-ui/react'
+'use client'
+import { Box, useDisclosure, VStack } from '@chakra-ui/react'
 import Cookies from 'js-cookie'
 import { useRouter } from 'next/router'
 import BottomDrawer from '../../shared/BottomDrawer'
 import { LeftBackRandom } from '../../shared/Headers/LeftBackRandom'
-import { SectionButton } from './SectionButton'
+import { SettingSectionList } from '@components/account/settings/SettingSectionList'
 
 export const SettingsTemplate = () => {
   const router = useRouter()
@@ -48,8 +49,8 @@ export const SettingsTemplate = () => {
 
   return (
     <Box h='full'>
+      Box
       <VStack bg='gray.100' h='full'>
-        {/* title */}
         <LeftBackRandom onClick={handleClickBack} title='설정' />
         {/* LOGOUT drawer */}
         <BottomDrawer
@@ -61,7 +62,6 @@ export const SettingsTemplate = () => {
           confirmLabel={'로그아웃'}
           onConfirm={handleLogout}
         />
-
         {/* SIGNOUT drawer */}
         <BottomDrawer
           headerLabel={'정말 비어랏을 떠나시는 건가요?'}
@@ -76,58 +76,8 @@ export const SettingsTemplate = () => {
           confirmLabel={'아뇨, 더 있을래요'}
           onConfirm={handleCancelSignout}
         />
-        <VStack
-          bg='gray.100'
-          pt='70px'
-          w='full'
-          h='full'
-          gap='10px'
-          borderRight={'1px solid'}
-          borderRightColor={'gray.200'}
-          borderLeft={'1px solid'}
-          borderLeftColor={'gray.200'}
-        >
-          <Flex w='full' flexDir={'column'}>
-            {NoticeSettingSection.map(({ title, href }) => (
-              <SectionButton key={title} title={title} href={href} />
-            ))}
-          </Flex>
-          <Flex w='full' h='full' flexDir={'column'}>
-            {usersSetting.map((item) => (
-              <SectionButton
-                title={item.title}
-                key={item.title}
-                onClick={item.onClick}
-                _hover={{}}
-              />
-            ))}
-            <Box
-              h='full'
-              bg={'white'}
-              w='full'
-              style={{
-                marginTop: 0,
-              }}
-            />
-          </Flex>
-        </VStack>
+        <SettingSectionList usersSetting={usersSetting} />
       </VStack>
     </Box>
   )
 }
-
-export const BeerSettingSection = [
-  { title: '최애맥주 변경', href: '/account/settings/favoritebeer' },
-]
-
-export const NoticeSettingSection = [
-  {
-    title: '공지사항',
-    href: '/account/settings/notice',
-  },
-  {
-    title: '문의하기',
-    href: '/account/settings/inquiry',
-  },
-  { title: '비어랏 정보', href: '/account/settings/info' },
-]
